pyproject.toml
[project]
name = "qdrant-neo4j-graphrag"
version = "0.1.0"
description = "GraphRAG client integrating Neo4j, Qdrant, LangChain, and LLMs."
authors = [{ name = "Devansh", email = "your-email@example.com" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "neo4j==5.28.1",
    "qdrant-client==1.14.3",
    "langchain-community==0.3.10",
    "pydantic==2.11.3"
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["uv", "setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
graphrag = "qdrant_neo4j_graphrag.__main__:main"
